requirements.txt
azure-devops
pandas
requests